+++ importer-creator/transform_script.py
@@ -4,16 +4,20 @@ from transform_logic import transform
 
 if __name__ == "__main__":
     import sys
+    import os
 
     inputfile = sys.argv[1]
     user_provided_context = sys.argv[2]
     verbose = sys.argv[3]
 
+    full_path = os.path.join("./files/", inputfile)
+
     print("inputfile: ", inputfile)
+    print("full_path: ", full_path)
     print("user_provided_context: ", user_provided_context)
     print("verbose: ", verbose)
 
-    generated_script = transform(inputfile, user_provided_context, verbose)
+    generated_script = transform(full_path, user_provided_context, verbose)
 
     # Write the generated script to a file
     with open("generated_script.py", "w") as file:
